The wonders of sleeping on ita.
I knew from the beginning that I wanted to stream rather than read the whole file into memory. But I made the mistake of thinking I could start with a string-based implementation and refactor.

Reading a stream changes the whole architecture. (With the string implementation, "architecture" was overkill.)

Oh well. The test cases are still valuable. "Build one to throw away" is as true now as it was in the 70s.